chore: reorganize project directory structure
Moved files and directories to create a more organized project structure.

- Updated file locations for better modularity and clarity.
- Adjusted import paths in affected files accordingly.
- No functional changes made, only structural adjustments.
