+++ FrontEnd/frontend/src/App.js
@@ -1,12 +1,7 @@
 import React from 'react';
-import AdminDashboard from './components/adminpage/admindashboard/AdminDashBoard';
-import { ToastContainer } from 'react-toastify';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'react-toastify/dist/ReactToastify.css';
-import BlogDetail from './components/adminpage/blogdetail/BlogDetail';
-import CreateBlog from './components/adminpage/createblog/CreateBlog';
-import Login from './components/login/Login';
-import SignUp from './pages/SignUp/SignUp';
+import Login from './pages/Login/Login';
 
 const App = () => {
   return (
