--- FrontEnd/frontend/src/components/adminpage/blogdetail/BlogDetail.jsx
@@ -1,128 +1,0 @@
-import React from "react";
-import "./BlogDetail.css";
-
-function BlogDetail() {
-    return (
-        <div className = "blogdetail_container">
-            <div className="div">
-                <div className="div-2">
-                <div className="div-3">
-                    <div className="div-4">Blog Review Film</div>
-                    <div classNameName="div-5">
-                    </div>
-                </div>
-                </div>
-                <div className="div-10">
-                <div className="div-11">
-                    <div className="column">
-                    <div className="div-12">
-                        <div className="div-13">Có nên quỳ gối trước khó khăn?</div>
-                        <img
-                        loading="lazy"
-                        srcset="https://cdn.builder.io/api/v1/image/assets/TEMP/6bab638981d46bad73cd6c20b3ca5e9d08e5b4d747f31150a1c7faa0fb896dec?apiKey=e4dc791f50af483d9388ff49ffbbba5e&width=100 100w, https://cdn.builder.io/api/v1/image/assets/TEMP/6bab638981d46bad73cd6c20b3ca5e9d08e5b4d747f31150a1c7faa0fb896dec?apiKey=e4dc791f50af483d9388ff49ffbbba5e&width=200 200w, https://cdn.builder.io/api/v1/image/assets/TEMP/6bab638981d46bad73cd6c20b3ca5e9d08e5b4d747f31150a1c7faa0fb896dec?apiKey=e4dc791f50af483d9388ff49ffbbba5e&width=400 400w, https://cdn.builder.io/api/v1/image/assets/TEMP/6bab638981d46bad73cd6c20b3ca5e9d08e5b4d747f31150a1c7faa0fb896dec?apiKey=e4dc791f50af483d9388ff49ffbbba5e&width=800 800w, https://cdn.builder.io/api/v1/image/assets/TEMP/6bab638981d46bad73cd6c20b3ca5e9d08e5b4d747f31150a1c7faa0fb896dec?apiKey=e4dc791f50af483d9388ff49ffbbba5e&width=1200 1200w, https://cdn.builder.io/api/v1/image/assets/TEMP/6bab638981d46bad73cd6c20b3ca5e9d08e5b4d747f31150a1c7faa0fb896dec?apiKey=e4dc791f50af483d9388ff49ffbbba5e&width=1600 1600w, https://cdn.builder.io/api/v1/image/assets/TEMP/6bab638981d46bad73cd6c20b3ca5e9d08e5b4d747f31150a1c7faa0fb896dec?apiKey=e4dc791f50af483d9388ff49ffbbba5e&width=2000 2000w, https://cdn.builder.io/api/v1/image/assets/TEMP/6bab638981d46bad73cd6c20b3ca5e9d08e5b4d747f31150a1c7faa0fb896dec?apiKey=e4dc791f50af483d9388ff49ffbbba5e&"
-                        className="img"
-                        />
-                    </div>
-                    </div>
-                    <div className="column-2">
-                    <div className="div-14">
-                        <div className="div-15">
-                        <div className="div-16">
-                            <div className="column-3">
-                            <div className="div-17">
-                                <div className="div-18">Post Date</div>
-                                <div className="div-19">January 9th, 2024</div>
-                                <div className="div-20">Catagory</div>
-                                <div className="div-21">Funny</div>
-                            </div>
-                            </div>
-                            <div className="column-4">
-                            <div className="div-22">
-                                <div className="div-23">Region</div>
-                                <div className="div-24">Korean</div>
-                                <div className="div-25">Score</div>
-                                <div className="div-26">9/10</div>
-                            </div>
-                            </div>
-                        </div>
-                        </div>
-                        <div className="div-27">
-                        Chỉ sau duy nhất 4 tập phim, Itaewon className (Tầng Lớp
-                        <br />
-                        Itaewon) đã khiến khán giả từ hả hê tới run rẩy vì
-                        <br />
-                        độ gắt đến từng tình tiết lẫn diễn xuất của dàn diễn
-                        <br />
-                        viên trẻ tài năng, tiêu biểu là &quot;phó chủ tịch&quot; Park
-                        <br />
-                        Seo Joon giờ đã trở thành một cựu tù nhân với hai
-                        <br />
-                        bàn tay trắng xây dựng cơ nghiệp.
-                        </div>
-                        <div className="div-28">
-                        Chẳng cần phải là twist mới đủ khiến người xem hoảng
-                        <br />
-                        loạn, Itaewon className đi từ hết bất ngờ này tới bất
-                        <br />
-                        ngờ khác với những hướng tình huống gây sang chấn
-                        <br />
-                        tinh thần xoay quanh sự phản bội, thân thế của từng
-                        <br />
-                        nhân vật lẫn thái độ của Park Sae Ro Yi đối với thực
-                        <br />
-                        tại cuộc sống nghiệt ngã. Với những phản hồi tích
-                        <br />
-                        cực từ cả khán giả Hàn lẫn Việt, Itaewon className có gì
-                        <br />
-                        thu hút đến như vậy?
-                        </div>
-                    </div>
-                    </div>
-                </div>
-                </div>
-                <div className="div-29">
-                <div className="div-30">
-                    <div className="div-31">Home</div>
-                    <div className="div-32">About us</div>
-                    <div className="div-33">Services</div>
-                    <div className="div-34">Terms</div>
-                    <div className="div-35">Privacy</div>
-                </div>
-                <div className="div-36">
-                    <div className="div-37">Blog</div>
-                    <div className="div-38">Recent Blog</div>
-                    <div className="div-39">List Blog</div>
-                    <div className="div-40">Top Film</div>
-                </div>
-                <div className="div-41">
-                    <div className="div-42">Reviewer</div>
-                    <div className="div-43">Sign in</div>
-                    <div className="div-44">List Reviewer</div>
-                    <div className="div-45">Top Reviewer</div>
-                    <div className="div-46">Privacy</div>
-                </div>
-                <div className="div-47">
-                    <div className="div-48">Contact</div>
-                    <div className="div-49">43 Hoa Thuan, Da Nang</div>
-                    <div className="div-50">+84-456-7890</div>
-                    <div className="div-51">+84-456-7890</div>
-                    <div className="div-52">java1@gmail.com</div>
-                </div>
-                </div>
-                <div className="div-53">
-                <div className="div-54">
-                    <div className="div-55">
-                    <div className="div-56"></div>
-                    <div className="div-57"></div>
-                    <div className="div-58"></div>
-                    <div className="div-59"></div>
-                    </div>
-                    <div className="div-60">Review Film</div>
-                </div>
-                </div>
-            </div>
-        </div>
-            
-    );
-}
-export default BlogDetail;
