--- FrontEnd/frontend/src/components/reviewerpage/reviewerFilm/FilmReviewer.jsx
@@ -1,12 +1,0 @@
-import ReviewerPage from "../../../pages/Reviewer/reviewer";
-
-const FilmReviewer = () => {
-
-    
-    return ( 
-    <>
-        <ReviewerPage />
-    </> );
-}
-
-export default FilmReviewer;
