--- FrontEnd/frontend/src/components/services/config.js
@@ -1,4 +1,0 @@
-// config.js
-const appURL = 'http://192.168.1.120:8080';
-
-export default appURL;
