+++ FrontEnd/frontend/src/configs/config.js
@@ -1,4 +1,4 @@
 // config.js
-const appURL = 'http://14.225.36.84:8080';
+const appURL = 'http://192.168.1.120:8080';
 
 export default appURL;
