+++ FrontEnd/frontend/src/index.js
@@ -1,18 +1,14 @@
-
 import React from"react";
 import ReactDOM from"react-dom/client";
 import App from"./App";
 import { BrowserRouter, Route, Routes } from"react-router-dom";
-import Login from"./components/login/Login";
-import AdminBlog from"./components/adminpage/adminblog/AdminBlog";
-import BlogDetail from "./components/adminpage/blogdetail/BlogDetail";
-import CreateBlog from "./components/adminpage/createblog/CreateBlog";
-import AdminDashBoard from "./components/adminpage/admindashboard/AdminDashBoard";
-import AdminFilm from "./components/adminpage/adminfilm/AdminFilm";
-import AdminAccount from "./components/adminpage/adminaccount/AdminAccount";
+
 import SignUp from "./pages/SignUp/SignUp";
-import DashBoardReviewer from "./components/reviewerpage/reviewerDashboard/DashboardReviewer";
-import FilmReviewer from "./components/reviewerpage/reviewerFilm/FilmReviewer";
+import Login from "./pages/Login/Login";
+import UserPage from "./pages/User/DashboardPage";
+import ReviewerPage from "./pages/User/ReviewerPage";
+import FilmPage from "./pages/User/FilmPage";
+import BlogPage from "./pages/User/BlogPage";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 
@@ -23,13 +19,10 @@ root.render(
         <Route index element={<App />} />
         <Route path="login" element={<Login />} />
         <Route path="signup" element={<SignUp />} />
-        <Route path="/admin/blog" element={<AdminBlog />} />
-        <Route path="/admin/edit" element={<BlogDetail />} />
-        <Route path="/admin/dashboard" element={<AdminDashBoard />} />
-        <Route path="/admin/film" element={<AdminFilm />} />
-        <Route path="/admin/account" element={<AdminAccount />} />
-        <Route path="/reviewer" element={<DashBoardReviewer />} />
-        <Route path="/reviewer/film" element={<FilmReviewer />} />
+        <Route path="/user" element={<UserPage />} />
+        <Route path="/user/reviewers" element={<ReviewerPage />} />
+        <Route path="/user/films" element={<FilmPage />} />
+        <Route path="/user/blogs" element={<BlogPage />} />
       </Routes>
     </BrowserRouter>
   </React.StrictMode>
