+++ FrontEnd/frontend/src/pages/User/BlogPage.jsx
@@ -1,0 +1,13 @@
+import Blog from "../../components/blog/Blog";
+import NavBarUser from "../../components/navbar/NavBarUser";
+
+const BlogPage = () => {
+    return ( 
+        <>
+            <NavBarUser />
+            <Blog />
+        </>
+     );
+}
+
+export default BlogPage;
