+++ FrontEnd/frontend/src/pages/User/DashboardPage.jsx
@@ -1,0 +1,14 @@
+import DashBoard from "../../components/dashboard/Dashboard";
+import NavBarUser from "../../components/navbar/NavBarUser";
+
+
+const DashBoardPage = () => {
+    return ( 
+        <>
+            <NavBarUser />
+            <DashBoard />
+        </>
+     );
+}
+
+export default DashBoardPage;
