+++ FrontEnd/frontend/src/pages/User/FilmPage.jsx
@@ -1,0 +1,13 @@
+import Film from "../../components/film/Film";
+import NavBarUser from "../../components/navbar/NavBarUser";
+
+function FilmPage() {
+    return ( 
+        <>
+            <NavBarUser />
+            <Film />
+        </>
+     );
+}
+
+export default FilmPage;
