+++ FrontEnd/frontend/src/pages/User/ReviewerPage.jsx
@@ -1,0 +1,13 @@
+import NavBarUser from "../../components/navbar/NavBarUser";
+import Reviewer from "../../components/reviewer/Reviewer";
+
+function ReviewerPage() {
+    return ( 
+        <>
+            <NavBarUser />
+            <Reviewer />
+        </>
+     );
+}
+
+export default ReviewerPage;
