+++ FrontEnd/frontend/src/services/ReviewerService.js
@@ -1,13 +1,13 @@
 import axios from "./customize-axios";
 
-  const fetchSummaryData = (accessToken) => {
-    const query = `/api/reviewer/dashboard`;
-    return axios.get(query, {
-        headers: {
-            'Authorization': `Bearer ${accessToken}`
-        }
-    });
+  // const fetchSummaryData = (accessToken) => {
+  //   const query = `/api/reviewer/dashboard`;
+  //   return axios.get(query, {
+  //       headers: {
+  //           'Authorization': `Bearer ${accessToken}`
+  //       }
+  //   });
    
-  };
+  // };
   
-  export { fetchSummaryData };
+  // export { fetchSummaryData };
